+++ lib/kitawi.dart
@@ -45,7 +45,7 @@ void render(Widget widget, Element? element) {
 /// The [VoidCallback] is called when the window is loaded and when the window
 /// is resized.
 void start(VoidCallback callback) {
-  /// The [VoidCallback] is called when the window is loaded.
+  // The [VoidCallback] is called when the window is loaded.
   Size().updateSize();
   Theme().updateMode();
   callback();