+++ lib/src/core/size.dart
@@ -6,18 +6,24 @@ class Size {
   late int _x;
   late int _y;
 
+  /// [Size] factory
   static final Size _instance = Size._internal();
 
+  /// [Size] accessor
   factory Size() {
     return _instance;
   }
 
+  /// [Size] constructor
   Size._internal() {
     _x = 0;
     _y = 0;
   }
 
+  /// The width of the window.
   double get width => _x.toDouble();
+
+  /// The height of the window.
   double get height => _y.toDouble();
 
   set x(int value) {
@@ -28,6 +34,7 @@ class Size {
     _y = value;
   }
 
+  /// Updates the size of the window.
   void updateSize() {
     _x = window.innerWidth ?? 0;
     _y = window.innerHeight ?? 0;