+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 0.0.1 (2023-02-21)
 
 - Here we go!
+
+## 0.0.2 (2023-02-22)
+
+- updated kitawi cli (kitawi create/run)
+- added basic widget support
+- minor state management improvements
+- migrated to webdev